refactor navbar using json and map
