+++ src/App.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "./index.css";
 import { Route, Routes } from "react-router-dom";
-import Navbar from "./components/Common/Navbar";
+import Navbar from "./components/Common/Navbar/Navbar";
 import ErrorPage from "./components/ErrorPage";
 import About from "./components/About/About";
 import Nonprofits from "./components/Apply/NonprofitsApply";
