+++ src/components/Home/Sections/Hero.jsx
@@ -21,7 +21,7 @@ function Hero() {
           id="bubble-repulse"
           className="flex justify-center flex-col items-center content-center"
         >
-          <img src={logo} className="w-96 px-10 md:max-w-sm" alt="Blueprint Logo" />
+          <img src={logo} className="w-72 px-10 md:max-w-sm" alt="Blueprint Logo" />
           <h2 className="pb-6 m-2 text-2xl font-bold">Technology for Social Good</h2>
           <div className="">
             <Link to="apply" type="button" className="blue-button-lg">
